Pixel data now stored as array of booleans
Each pixel was previously represented by an object with an "on" boolean
and an id. This is unnecessarily complicated for most cases, and it made
it harder to avoid directly manipulating component state. Now, if I want
to manipulate the pixel data, just slice the array held in state and I
have a fresh copy that can be manipulated and passed to setState when
ready.