Added little arrows in the mockup container
We're creating these arrows by drawing to a canvas. This allows us to
keep the scale flexible, as opposed to using static images.

To do this, adding general functions for creating a canvas with a
particular image out of pixel data. These functions are also useful in
generating our "tile" image out of the pixel data when we show our
mockup image, and when we set the background.